Separate poll view into new files
