+++ src/components/create-poll.jsx
@@ -30,11 +30,6 @@ import {
   KeyboardDatePicker,
   MuiPickersUtilsProvider,
 } from '@material-ui/pickers';
-import {
-  ptBR as pt_brLocale,
-  es as es_Locale,
-  enUS as en_usLocale,
-} from 'date-fns/locale';
 
 // Constantes
 import { POLL_TYPES } from '@/utils/constants';
@@ -131,20 +126,6 @@ function CreatePoll({ open, onClose, intl }) {
   const [newOptionError, setNewOptionError] = useState(false);
 
   const languageContext = useContext(LocaleContext);
-  const [dateLocale, setDateLocale] = useState(pt_brLocale);
-
-  useEffect(() => {
-    switch (languageContext.locale) {
-      case 'pt-BR':
-        setDateLocale(pt_brLocale);
-        break;
-      case 'es-ES':
-        setDateLocale(es_Locale);
-        break;
-      default:
-        setDateLocale(en_usLocale);
-    }
-  }, [languageContext]);
 
   // Estado de erro de preenchimento da deadline
   const deadlineError = useMemo(() => {
@@ -384,7 +365,10 @@ function CreatePoll({ open, onClose, intl }) {
                   ))}
                 </Select>
               </Grid>
-              <MuiPickersUtilsProvider utils={DateFnsUtils} locale={dateLocale}>
+              <MuiPickersUtilsProvider
+                utils={DateFnsUtils}
+                locale={languageContext.dateLocale}
+              >
                 <Grid item xs={12}>
                   <InputLabel htmlFor='deadline'>
                     {intl.formatMessage(messages.deadline)}
@@ -408,7 +392,7 @@ function CreatePoll({ open, onClose, intl }) {
                       languageContext.locale === 'pt-BR' ||
                       languageContext.locale === 'es-ES'
                         ? 'dd/MM/yyyy'
-                        : 'yyyy/MM/dd'
+                        : 'MM/dd/yyyy'
                     }
                   />
                 </Grid>
