+++ src/messages/en.json
@@ -41,9 +41,9 @@
   "confirm-dialog.question": "Are you sure?",
   "confirm-dialog.yes": "Yes, continue",
   "confirm-dialog.no": "No, cancel",
-  "manage.poll.individual-email-sent": "Email sent to ",
+  "manage.poll.individual-email-sent": "Email sent to {email}",
   "manage.poll.send-invidual-email": "Send individual email",
-  "manage.poll.added-emails-success": " emails were added!",
+  "manage.poll.added-emails-success": "{count, plural, =1 {One email was} other {# emails were}} added!",
   "manage.poll.add-emails": "Add e-mails:",
   "manage.poll.email-already-exists": "E-mail already registered",
   "manage.add": "Add",
