+++ src/messages/es.json
@@ -41,9 +41,9 @@
   "confirm-dialog.question": "¿Estás seguro?",
   "confirm-dialog.yes": "Si continuar",
   "confirm-dialog.no": "No, cancelar",
-  "manage.poll.individual-email-sent": "Correo electrónico enviado a ",
+  "manage.poll.individual-email-sent": "Correo electrónico enviado a {email}",
   "manage.poll.send-invidual-email": "Enviar correo electrónico individual",
-  "manage.poll.added-emails-success": " correos electrónicos se han agregado",
+  "manage.poll.added-emails-success": "{count, plural, =1 {Se agregó un correo electrónico} other {Se agregaron # correos electrónicos}}",
   "manage.poll.add-emails": "Agregar correos electrónicos:",
   "manage.poll.email-already-exists": "Correo electrónico ya registrado",
   "manage.add": "Agregar",
@@ -55,7 +55,7 @@
   "manage.delete": "Eliminar",
   "manage.back": "Volver",
   "manage.poll.emails-successfully-sent": "Envió correos electrónicos!",
-  "manage.poll.result": "Resultados de las elecciones",
+  "manage.poll.result": "Resultado de la elección",
   "manage.poll.votes": "votos",
   "manage.poll.send-links": "Envíe enlaces de votación a todos los inscritos",
   "manage.poll.group": "Grupo",
