+++ src/messages/pt.json
@@ -41,9 +41,9 @@
   "confirm-dialog.question": "Você tem certeza?",
   "confirm-dialog.yes": "Sim, continuar",
   "confirm-dialog.no": "Não, cancelar",
-  "manage.poll.individual-email-sent": "Email enviado para ",
+  "manage.poll.individual-email-sent": "Email enviado para {email}",
   "manage.poll.send-invidual-email": "Enviar e-mail individual",
-  "manage.poll.added-emails-success": " emails foram adicionados!",
+  "manage.poll.added-emails-success": "{count, plural, =1 {Um email foi adicionado!} other {# emails foram adicionados!}}",
   "manage.poll.add-emails": "Adicionar e-mails:",
   "manage.poll.email-already-exists": "Email já cadastrado",
   "manage.add": "Adicionar",
